requirements.txt
# Web framework
Flask==3.0.3
Werkzeug==3.0.3

# Database
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10      # PostgreSQL driver

# HTTP client (MobSF API calls)
requests==2.32.3

# Environment variables
python-dotenv==1.0.1

# Data analysis (comparator module)
pandas==2.2.3
numpy==2.1.3

# Production WSGI server
gunicorn==23.0.0
